Rename AppSec::ActionHandler to AppSec::ActionsHandler
Because we need to handle actions according their precedence.
