+++ lib/datadog/appsec/contrib/sinatra/gateway/watcher.rb
@@ -43,9 +43,7 @@ module Datadog
                       Datadog::AppSec::Event.tag_and_keep!(context, result)
                       context.events << event
 
-                      result.actions.each do |action_type, action_params|
-                        Datadog::AppSec::ActionHandler.handle(action_type, action_params)
-                      end
+                      Datadog::AppSec::ActionsHandler.handle(result.actions)
                     end
                   end
 
@@ -77,9 +75,7 @@ module Datadog
                       Datadog::AppSec::Event.tag_and_keep!(context, result)
                       context.events << event
 
-                      result.actions.each do |action_type, action_params|
-                        Datadog::AppSec::ActionHandler.handle(action_type, action_params)
-                      end
+                      Datadog::AppSec::ActionsHandler.handle(result.actions)
                     end
                   end
